wrapper/liquidity-helper: Upgrade code for db migration
Code for updating reward claim requests in wrapper's tables to the
correct to/to-guard destinations.

Code for transferring funds out of the liquidity-helper's temporary
account during the upgrade.

The full transfer and update list will be in the transaction data, but
will be sourced from on-chain data via a small script that will get
added in a follow-up commit.